Refactor class and @staticmethod to simple fn
- Rename cunao2.py to cunao.py
